cleanup
* move utility functions to typeutils
* remove unused import
* change docstring to imperative
