+++ andi/typeutils.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
-from typing import Union, List
+import sys
+from typing import Union, List, Callable, Dict
 
 
 def is_union(tp) -> bool:
@@ -43,3 +44,44 @@ def issubclass_safe(cls, bases) -> bool:
         return issubclass(cls, bases)
     except TypeError:
         return False
+
+
+def get_globalns(func: Callable) -> Dict:
+    """ Return the global namespace that will be used for the resolution
+    of postponed type annotations.
+
+    Based on ``typing.get_type_hints`` code, with a workaround for ``attrs``
+    issue.
+    """
+    ns = dict(_get_globalns_for_attrs(func))
+    ns.update(_get_globalns_as_get_type_hints(func))
+    return ns
+
+
+def _get_globalns_as_get_type_hints(func: Callable) -> Dict:
+    """ Global namespace resolution extracted from ``get_type_hints`` method.
+    Python 3.7 (https://github.com/python/cpython/blob/3.7/Lib/typing.py#L981-L988)
+    Note that this is only supporting functions as input. """
+    nsobj = func
+    # Find globalns for the unwrapped object.
+    while hasattr(nsobj, '__wrapped__'):
+        nsobj = getattr(nsobj, '__wrapped__')
+    return getattr(nsobj, '__globals__', {})
+
+
+def _get_globalns_for_attrs(func: Callable) -> Dict:
+    """ Adds partial support for postponed type annotations in attrs classes.
+    Also required to support attrs classes when
+    ``from __future__ import annotations`` is used (default for python 4.0).
+    See https://github.com/python-attrs/attrs/issues/593 """
+    if func.__module__ in sys.modules:
+        return dict(sys.modules[func.__module__].__dict__)
+    else:
+        # Theoretically this can happen if someone writes
+        # a custom string to func.__module__.  In which case
+        # such attrs might not be fully introspectable
+        # (w.r.t. typing.get_type_hints) but will still function
+        # correctly.
+        return {}
+
+
